+++ app.vue
@@ -8,7 +8,4 @@
 
 <style lang="scss">
 @use "~/assets/global.scss";
-.overflow-hidden {
-  overflow: hidden;
-}
 </style>
