+++ assets/global.scss
@@ -1,0 +1,5 @@
+// este estilo ya existe como .lista-sin-estilo en sisdai-css
+ul.lista-sin-dibujo {
+  list-style: none;
+  padding-left: 0;
+}
