+++ components/consulta/ElementoSeleccionado.vue
@@ -1,5 +1,6 @@
 <script setup>
-const selectedStore = useSelectedResourcesStore();
+import { useSelectedResourcesStore } from "~/stores/selectedResources.js";
+const resourcesStore = useSelectedResourcesStore();
 
 const props = defineProps({
   selectedElement: {
@@ -9,77 +10,75 @@ const props = defineProps({
   resourceType: { type: String, required: true },
 });
 const { selectedElement, resourceType } = toRefs(props);
-
-if (!selectedStore.shownFiles[resourceType.value]) {
-  //console.log("si vacio", !selectedStore.shownFiles[resourceType.value]);
-  let firstSelection = selectedStore.selectedResources[resourceType.value][0];
-  selectedStore.setShownFile(resourceType.value, firstSelection);
-}
-
-/* const isChecked = computed(() => {
-  if (resourceType.value !== "dataLayer") {
-    if (
-      selectedElement.value.uuid ===
-      selectedStore.shownFiles[resourceType.value].uuid
-    ) {
-      return true;
-    } else {
-      return false;
-    }
-  } else {
-    return true;
-  }
-}); */
 </script>
 <template>
-  <div class="tarjeta m-y-1">
-    <!--     <div :class="isChecked ? 'tarjeta-selected' : 'selected-unselected'"> -->
-    <div class="tarjeta-selected">
-      <div class="flex flex-contenido-separado m-0 encabezado-tarjeta">
-        <p class="tarjeta-texto-secundario m-0">Categoria</p>
-        <div class="m-0">
-          <span
-            class="pictograma-informacion pictograma-mediano"
-            aria-hidden="true"
-          ></span>
-          <span
-            class="pictograma-subir-capa pictograma-mediano"
-            aria-hidden="true"
-            @click="selectedStore.raiseIndex(selectedElement, resourceType)"
-          ></span>
-          <span
-            class="pictograma-bajar-capa pictograma-mediano"
-            aria-hidden="true"
-            @click="selectedStore.lowerIndex(selectedElement, resourceType)"
-          ></span>
-        </div>
+  <div class="tarjeta-seleccionada fondo-color-acento">
+    <div class="header-tarjeta">
+      <p>Categoria</p>
+      <div>
+        <span
+          class="pictograma-informacion pictograma-chico"
+          aria-hidden="true"
+        ></span>
+        <span
+          class="pictograma-subir-capa pictograma-chico"
+          aria-hidden="true"
+        ></span>
+        <span
+          class="pictograma-bajar-capa pictograma-chico"
+          aria-hidden="true"
+        ></span>
       </div>
+    </div>
 
-      <ConsultaContenidoCapaSeleccionada
-        v-if="resourceType === 'dataLayer'"
-        :resource-type="resourceType"
-        :selected-element="selectedElement"
-      />
+    <div>
+      <h3>
+        {{ selectedElement.title }}
+      </h3>
 
-      <ConsultaContenidoDocSeleccionado
-        v-if="resourceType !== 'dataLayer'"
-        :group-name="resourceType"
-        :resource-type="resourceType"
-        :selected-element="selectedElement"
-      />
+      <div class="contenedor-botones">
+        <button
+          class="boton-pictograma"
+          aria-label="Remover selección"
+          type="button"
+          @click="resourcesStore.removeResource(resourceType, selectedElement)"
+        >
+          <span class="pictograma-eliminar" aria-hidden="true"></span>
+        </button>
+        <button
+          class="boton-pictograma"
+          aria-label="Descargar selección"
+          type="button"
+        >
+          <span class="pictograma-archivo-descargar" aria-hidden="true"></span>
+        </button>
+      </div>
     </div>
   </div>
 </template>
 <style lang="scss" scoped>
-.tarjeta-selected {
-  padding: 16px;
-  background-color: var(--color-secundario-2);
+.tarjeta-seleccionada {
+  margin: 15px 10px;
+  h3 {
+    margin: 4px;
+  }
 }
-.selected-unselected {
-  padding: 16px;
-  background-color: var(--color-neutro-1);
-}
-.encabezado-tarjeta {
+.header-tarjeta {
+  display: flex;
+  flex-wrap: wrap;
   align-items: center;
+  justify-content: space-between;
+  p {
+    margin: 0px;
+  }
+  span {
+    margin: 0px;
+  }
+}
+
+.contenedor-botones {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: flex-end;
 }
 </style>
