+++ components/consulta/LayoutCatalogo.vue
@@ -1,6 +1,4 @@
 <script setup>
-// PENDING: El buscador sisdai genera un error que dice "Hidration completed but contains mismatches"
-const resourcesStore = useSelectedResourcesStore();
 const props = defineProps({
   titulo: { type: String, default: "Título" },
   resourceType: { type: String, required: true },
@@ -8,140 +6,63 @@ const props = defineProps({
 });
 const { titulo, resourceType } = toRefs(props);
 const { resourcesList } = useGeonodeResources({
+  pageNumber: 1,
+  pageSize: 20,
   resourceType: resourceType.value,
 });
-
-const filteredResources = ref([]);
-const config = useRuntimeConfig();
-const apiCategorias = `${config.public.geonodeApi}/facets/category`;
-const categoryList = ref([]);
-const categorizedResources = ref({});
-const selectedCategories = ref([]);
-
-// Esta parte es para obtener todas las categorias
-const { data: geonodeCategories } = await useFetch(`${apiCategorias}`);
-if (!geonodeCategories.value) {
-  categoryList.value = ["Sin clasificar"];
-} else {
-  geonodeCategories.value.topics.items.map((d) => {
-    categoryList.value.push(d.label);
-  });
-}
+const categories = ref({});
 
 function groupResults() {
-  categorizedResources.value = {};
-  filteredResources.value.map((r) => {
-    if (r.category) {
-      let title = r.category.gn_description;
-      if (Object.keys(categorizedResources.value).includes(title)) {
-        categorizedResources.value[title].push(r);
-      } else {
-        categorizedResources.value[title] = [];
-        categorizedResources.value[title].push(r);
-      }
+  categories.value = {};
+  resourcesList.value.map((r) => {
+    const anio = r.date.slice(0, 4);
+    if (Object.keys(categories.value).includes(anio)) {
+      categories.value[anio].push(r);
     } else {
-      if (Object.keys(categorizedResources.value).includes("Sin clasificar")) {
-        categorizedResources.value["Sin clasificar"].push(r);
-      } else {
-        categorizedResources.value["Sin clasificar"] = [];
-        categorizedResources.value["Sin clasificar"].push(r);
-      }
+      categories.value[anio] = [];
+      categories.value[anio].push(r);
     }
   });
 }
 
-function setSelectedCategory(categoria) {
-  if (selectedCategories.value.includes(categoria)) {
-    selectedCategories.value = selectedCategories.value.filter(
-      (c) => c !== categoria
-    );
-  } else {
-    selectedCategories.value.push(categoria);
-  }
-}
-
-// Se utiliza un watcher porque inicialmente resourcesList y filteredResources están vacías
+// Se utiliza un watcher porque inicialmente resourcesList está vacía
 watch(resourcesList, () => {
-  resourcesStore.updateFilteredResources(
-    resourceType.value,
-    resourcesList.value
-  );
+  groupResults();
 });
-watch(
-  () => resourcesStore.filteredResources[resourceType.value],
-  () => {
-    filteredResources.value =
-      resourcesStore.filteredResources[resourceType.value];
-    groupResults();
-  },
-  { deep: true }
-);
 </script>
 
 <template>
-  <div class="catalogo-layout">
-    <div class="controles-catalogo">
-      <p class="h4 fondo-color-acento p-3 m-0">{{ titulo }}</p>
+  <div>
+    <p class="h4 fondo-color-acento p-3 m-0">{{ titulo }}</p>
 
-      <div class="m-x-2 m-y-1">
-        <p class="m-0">Explora conjuntos de datos abiertos nacionales.</p>
-        <ClientOnly>
-          <ConsultaElementoBuscador
-            :resources-list="resourcesList"
-            :resource-type="resourceType"
-            :categories="categoryList"
-          />
-        </ClientOnly>
+    <div class="m-x-2 m-y-1">
+      <p class="m-0">Explora conjuntos de datos abiertos nacionales.</p>
 
-        <UiNumeroElementos
-          :numero="filteredResources.length"
-          :etiqueta="etiquetaElementos"
-        />
-      </div>
+      <ConsultaElementoBuscador />
+
+      <UiNumeroElementos
+        :numero="resourcesList.length"
+        :etiqueta="etiquetaElementos"
+      />
     </div>
 
     <div
-      v-for="category in Object.keys(categorizedResources)"
+      v-for="category in Object.keys(categories)"
       :key="category"
-      class="m-y-1"
+      class="panel-catalogo"
     >
-      <div class="">
-        <ConsultaElementoCategoria
-          :title="category"
-          :tag="etiquetaElementos"
-          :number-elements="categorizedResources[category].length"
-          @click="setSelectedCategory(category)"
-        />
-      </div>
-      <div
-        class="contenedor-archivos"
-        v-if="selectedCategories.includes(category)"
-        v-for="(option, index) in categorizedResources[category]"
-      >
-        <ConsultaElementoCatalogo
-          class="elemento-catalogo"
-          :key="index"
-          :catalogue-element="option"
-          :resource-type="resourceType"
-        />
-      </div>
+      <h2>{{ category }}</h2>
+      <ConsultaElementoCatalogo
+        v-for="(option, index) in categories[category]"
+        :key="index"
+        :catalogue-element="option"
+        :resource-type="resourceType"
+      />
     </div>
   </div>
 </template>
-<style lang="scss" scoped>
-.contenedor-archivos {
-  border-bottom: solid 2px var(--color-neutro-3);
-}
-.catalogo-layout {
-  height: 100vh;
-  overflow-x: hidden;
-  overflow-y: auto;
-}
-.controles-catalogo {
-  position: sticky;
-  top: 0;
-  z-index: 1;
-  background-color: var(--color-neutro-0);
-  padding-bottom: 8px;
+<style lang="scss">
+.panel-catalogo {
+  padding: 0px 8px;
 }
 </style>
