+++ pages/consulta.vue
@@ -1,32 +1,10 @@
 <script setup>
 definePageMeta({ middleware: "redireccionar-modulo-consulta" });
-
-const ruta = "/consulta";
-
-const storeConsulta = useConsultaStore();
 </script>
 
 <template>
   <div class="modulo-consultas flex">
-    <UiNavegacionLateral
-      :sub-paginas="[
-        {
-          pictograma: 'pictograma-capas',
-          ruta: `${ruta}/capas`,
-        },
-        {
-          pictograma: 'pictograma-tabla',
-          ruta: `${ruta}/tablas`,
-        },
-        {
-          pictograma: 'pictograma-documento',
-          ruta: `${ruta}/documentos`,
-        },
-      ]"
-      :funcion-colapsar="storeConsulta.alternarCatalogoColapsable"
-      :estado-colapable="storeConsulta.catalogoColapsado"
-      :idColapsable="storeConsulta.idNavegacionLateral"
-    />
+    <ConsultaNavegacionLateral />
 
     <div class="contenedor-contenido">
       <NuxtPage />
