+++ pages/consulta/tablas.vue
@@ -7,8 +7,6 @@ const tamanioPagina = 10;
 
 const config = useRuntimeConfig();
 
-const resourceType = "dataTable";
-
 const obtenerDatos = async () => {
   const url = new URL(`${config.public.geoserverUrl}/ows`);
   url.search = new URLSearchParams({
@@ -38,13 +36,7 @@ watch(paginaActual, obtenerDatos, { immediate: true });
 
 <template>
   <ConsultaLayoutPaneles>
-    <template #catalogo>
-      <ConsultaLayoutCatalogo
-        titulo="Tabulados de datos"
-        :resource-type="resourceType"
-        etiqueta-elementos="Datos tabulados"
-      />
-    </template>
+    <template #catalogo>Tablas disponibles</template>
 
     <template #visualizador>
       <UiTablaAccesible
@@ -58,12 +50,6 @@ watch(paginaActual, obtenerDatos, { immediate: true });
       />
     </template>
 
-    <template #seleccion>
-      <ConsultaLayoutSeleccion
-        titulo="Tabulados de datos"
-        :resource-type="resourceType"
-        etiqueta-elementos="Datos tabulaos"
-      />
-    </template>
+    <template #seleccion>Tablas seleccionadas</template>
   </ConsultaLayoutPaneles>
 </template>
