+++ pages/ia/index.vue
@@ -1,0 +1,6 @@
+<script setup>
+import VistaInicioChat from "~/components/ia/VistaInicioChat.vue";
+</script>
+<template>
+  <VistaInicioChat />
+</template>
