--- pages/ia/proyecto/_index.vue
@@ -1,4 +1,0 @@
-<script setup></script>
-<template>
-  <div />
-</template>
